+++ tests/testthat/test-paste_linter.R
@@ -8,6 +8,7 @@ test_that("paste_linter skips allowed usages for sep=''", {
   expect_no_lint("sep <- ''; paste('a', sep)", linter)
   expect_no_lint("paste(sep = ',', '', 'a')", linter)
   expect_no_lint("paste0('a', 'b', 'c')", linter)
+  expect_no_lint("expression(paste('a', b, sep = ''))", linter)
 })
 
 test_that("paste_linter blocks simple disallowed usages for sep=''", {
@@ -19,6 +20,12 @@ test_that("paste_linter blocks simple disallowed usages for sep=''", {
 
   expect_lint(
     "paste('a', 'b', sep = '')",
+    rex::rex('paste0(...) is better than paste(..., sep = "").'),
+    paste_linter()
+  )
+
+  expect_lint(
+    "paste('a', 'b', expression(2), sep = '')",
     rex::rex('paste0(...) is better than paste(..., sep = "").'),
     paste_linter()
   )